implement the func and small refactor
